+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -119,10 +119,7 @@ export function withExtendedResolver(
   // this needs to be unified since you can't dynamically
   // swap out the transformer based on platform.
   const assetRegistryPath = fs.realpathSync(
-    // This is the native asset registry alias for native.
-    path.resolve(resolveFrom(config.projectRoot, 'react-native/Libraries/Image/AssetRegistry'))
-    // NOTE(EvanBacon): This is the newer import but it doesn't work in the expo/expo monorepo.
-    // path.resolve(resolveFrom(projectRoot, '@react-native/assets-registry/registry.js'))
+    path.resolve(resolveFrom(config.projectRoot, '@react-native/assets-registry/registry.js'))
   );
 
   const defaultResolver = metroResolver.resolve;